feat(entities)!: move `apis_entites.abc` classes to `entities.abc`
BREAKING-CHANGE: We move the base classes from `apis_entities.abc` to
`entities.abc`. This affects
* `E21_Person`
* `E53_Place`
* `E74_Group`
* `SimpleLabelModel`
* All the rdf config classes in `apis_entities.rdfconfigs` are now
  located in `entities.rdfconfigs`
* All the templates referring to `E53_Place`
Anyone inheriting from those classes has to update their imports.
