+++ apis_core/apis_entities/locale/de/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2026-05-05 04:07-0500\n"
+"POT-Creation-Date: 2026-06-15 00:56-0500\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -18,74 +18,6 @@ msgstr ""
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1);\n"
 
-#: apis_core/apis_entities/abc.py
-msgid "forename"
-msgstr "Vorname"
-
-#: apis_core/apis_entities/abc.py
-msgid "surname"
-msgstr "Nachname"
-
-#: apis_core/apis_entities/abc.py
-msgid "gender"
-msgstr "Gender"
-
-#: apis_core/apis_entities/abc.py
-msgid "date of birth"
-msgstr "Geburtsdatum"
-
-#: apis_core/apis_entities/abc.py
-msgid "date of death"
-msgstr "Sterbedatum"
-
-#: apis_core/apis_entities/abc.py
-msgid "person"
-msgstr "Person"
-
-#: apis_core/apis_entities/abc.py
-msgid "persons"
-msgstr "Personen"
-
-#: apis_core/apis_entities/abc.py
-msgid "No name"
-msgstr "Kein Name"
-
-#: apis_core/apis_entities/abc.py
-msgid "label"
-msgstr "Bezeichnung"
-
-#: apis_core/apis_entities/abc.py
-msgid "latitude"
-msgstr "Breitengrad"
-
-#: apis_core/apis_entities/abc.py
-msgid "longitude"
-msgstr "Längengrad"
-
-#: apis_core/apis_entities/abc.py
-msgid "feature code"
-msgstr "Feature-Code"
-
-#: apis_core/apis_entities/abc.py
-msgid "place"
-msgstr "Ort"
-
-#: apis_core/apis_entities/abc.py
-msgid "places"
-msgstr "Orte"
-
-#: apis_core/apis_entities/abc.py
-msgid "No label"
-msgstr "Keine Bezeichnung"
-
-#: apis_core/apis_entities/abc.py
-msgid "group"
-msgstr "Gruppe"
-
-#: apis_core/apis_entities/abc.py
-msgid "groups"
-msgstr "Gruppen"
-
 #: apis_core/apis_entities/forms.py
 msgid "Search places"
 msgstr "Orte suchen"
@@ -93,14 +25,6 @@ msgstr "Orte suchen"
 #: apis_core/apis_entities/forms.py
 msgid "Click on a result to fill the form"
 msgstr "Ergebnis auswählen um die Daten zu übernehmen"
-
-#: apis_core/apis_entities/templates/apis_core/apis_entities/e53_place_autocomplete_result.html
-msgid "local result"
-msgstr "Lokaler Eintrag"
-
-#: apis_core/apis_entities/templates/apis_core/apis_entities/e53_place_list.html
-msgid "Map"
-msgstr "Karte"
 
 #: apis_core/apis_entities/templates/apis_entities/widgets/placelookup.html
 msgid "Search"
