+++ apis_core/entities/rdfconfigs/group.py
