+++ apis_core/entities/rdfconfigs/person.py
