+++ apis_core/entities/rdfconfigs/place.py
