+++ apis_core/entities/static/js/E53_Place_popover.js
