+++ apis_core/entities/templates/entities/e53_place_list.html
